requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
